requirements_dev.txt
pre-commit
mypy
ipykernel
pydantic
sqlalchemy
pyathena
fastapi
uvicorn
mangum
requests
mysql-connector-python
git+https://github.com/Oxford-Data-Processes/aws-utils.git
